fix require changes
